Commit of JS file (I messed up on the last one)
Messed up on the last commit and sent in the CSS and JS file at the
same time. On that last commit, you’ll see that I changed my update
function by adding in the code I got off of Stack Overflow. It told me
to past the code in right before the actual drawing occurs.